+++ examples/sglang_multiturn/config/gsm8k_multiturn_grpo.yaml
@@ -15,7 +15,7 @@ data:
 actor_rollout_ref:
   hybrid_engine: True
   rollout:
-    name: sglang_async
+    name: sglang
     multi_turn:
       enable: True
       max_turns: 5
