+++ tests/workers/rollout/test_async_sglang_server.py
@@ -1,3 +1,4 @@
+# Copyright 2023-2024 SGLang Team
 # Copyright 2025 Bytedance Ltd. and/or its affiliates
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -20,7 +21,6 @@ from omegaconf import DictConfig
 @patch.dict(
     "sys.modules",
     {
-        "verl.workers.rollout.sglang_rollout.async_sglang_rollout": MagicMock(AsyncSGLangRollout=MagicMock()),
         "verl.workers.rollout.sglang_rollout.sglang_rollout": MagicMock(SGLangRollout=MagicMock()),
     },
 )
