+++ tests/workers/rollout/test_sglang_async_rollout_search_tools.py
@@ -32,7 +32,7 @@ from verl.protocol import DataProto
 from verl.tools.schemas import OpenAIFunctionParametersSchema, OpenAIFunctionPropertySchema, OpenAIFunctionSchema, OpenAIFunctionToolSchema
 from verl.tools.search_tool import SearchTool
 from verl.workers.rollout.schemas import AsyncRolloutRequest, AsyncRolloutRequestStateEnum, Message
-from verl.workers.rollout.sglang_rollout.async_sglang_rollout import AsyncSGLangRollout
+from verl.workers.rollout.sglang_rollout.sglang_rollout import SGLangRollout
 
 DEFAULT_USER_CONTENT_PREFIX = (
     "Answer the given question. You must conduct reasoning inside <think> and </think> "
@@ -143,11 +143,11 @@ class TestRolloutWithSearchTools:
         prompts = DataProto(batch=prompt_dict, non_tensor_batch={"raw_prompt": messages, "tools_kwargs": tools_kwargs, "index": index})
         return prompts
 
-    @patch.object(AsyncSGLangRollout, "_init_distributed_env", return_value=None)
-    @patch.object(AsyncSGLangRollout, "_init_inference_engine", return_value=None)
-    @patch.object(AsyncSGLangRollout, "_init_sampling_params", return_value=None)
+    @patch.object(SGLangRollout, "_init_distributed_env", return_value=None)
+    @patch.object(SGLangRollout, "_init_inference_engine", return_value=None)
+    @patch.object(SGLangRollout, "_init_sampling_params", return_value=None)
     def test_tools_registration(self, mock_env, mock_engine, mock_sampling, search_rollout_config, qwen_tokenizer, qwen_model_config):
-        rollout = AsyncSGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
+        rollout = SGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
         assert len(rollout._tool_schemas) == 1
         assert "search" in rollout._tool_map.keys()
         from verl.tools.search_tool import SearchTool
@@ -156,11 +156,11 @@ class TestRolloutWithSearchTools:
         # depend on the tokenizer
         assert rollout._tool_call_parser_type == "qwen25"
 
-    @patch.object(AsyncSGLangRollout, "_init_distributed_env", return_value=None)
-    @patch.object(AsyncSGLangRollout, "_init_inference_engine", return_value=None)
-    @patch.object(AsyncSGLangRollout, "_init_sampling_params", return_value=None)
+    @patch.object(SGLangRollout, "_init_distributed_env", return_value=None)
+    @patch.object(SGLangRollout, "_init_inference_engine", return_value=None)
+    @patch.object(SGLangRollout, "_init_sampling_params", return_value=None)
     def test_rollout_req_creation(self, mock_env, mock_engine, mock_sampling, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto):
-        rollout = AsyncSGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
+        rollout = SGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
         req_list = rollout._preprocess_prompt_to_async_rollout_requests(search_data_proto, n=1)
         assert len(req_list) == 1
         assert req_list[0].state == AsyncRolloutRequestStateEnum.PENDING
@@ -186,12 +186,12 @@ class TestRolloutWithSearchTools:
             ),
         )
 
-    @patch.object(AsyncSGLangRollout, "_init_distributed_env", return_value=None)
-    @patch.object(AsyncSGLangRollout, "_init_inference_engine", return_value=None)
-    @patch.object(AsyncSGLangRollout, "_init_sampling_params", return_value=None)
+    @patch.object(SGLangRollout, "_init_distributed_env", return_value=None)
+    @patch.object(SGLangRollout, "_init_inference_engine", return_value=None)
+    @patch.object(SGLangRollout, "_init_sampling_params", return_value=None)
     def test_over_size_case(self, mock_env, mock_engine, mock_sampling, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto, search_data):
         search_rollout_config.multi_turn.max_turns = 1
-        rollout = AsyncSGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
+        rollout = SGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
         req = rollout._preprocess_prompt_to_async_rollout_requests(search_data_proto, n=1)[0]
         req = MagicMock(wraps=req, spec=AsyncRolloutRequest)
         req.finalize = MagicMock()
@@ -223,9 +223,9 @@ class TestRolloutWithSearchTools:
         )
 
     @patch.object(SearchTool, "execute", new_callable=AsyncMock)
-    @patch.object(AsyncSGLangRollout, "_init_distributed_env", return_value=None)
-    @patch.object(AsyncSGLangRollout, "_init_inference_engine", return_value=None)
-    @patch.object(AsyncSGLangRollout, "_init_sampling_params", return_value=None)
+    @patch.object(SGLangRollout, "_init_distributed_env", return_value=None)
+    @patch.object(SGLangRollout, "_init_inference_engine", return_value=None)
+    @patch.object(SGLangRollout, "_init_sampling_params", return_value=None)
     def test_tool_call_basic_case(self, mock_sampling, mock_engine, mock_env, mock_execute, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto, search_data):
         _, expect_turn_array, tool_return_array = search_data
 
@@ -233,7 +233,7 @@ class TestRolloutWithSearchTools:
         mock_execute.side_effect = [(msg, 0.0, {"status": "success"}) for msg in tool_return_array]
 
         search_rollout_config.multi_turn.max_turns = 10
-        rollout = AsyncSGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
+        rollout = SGLangRollout(actor_module="", config=search_rollout_config, tokenizer=qwen_tokenizer, model_hf_config=qwen_model_config)
 
         rollout._tool_map["search"].retrieval_service_url = "mock://dummy"
 
@@ -272,9 +272,9 @@ class TestRolloutWithSearchTools:
         assert search_counter == 2
 
     @patch.object(SearchTool, "execute", new_callable=AsyncMock)
-    @patch.object(AsyncSGLangRollout, "_init_distributed_env", return_value=None)
-    @patch.object(AsyncSGLangRollout, "_init_inference_engine", return_value=None)
-    @patch.object(AsyncSGLangRollout, "_init_sampling_params", return_value=None)
+    @patch.object(SGLangRollout, "_init_distributed_env", return_value=None)
+    @patch.object(SGLangRollout, "_init_inference_engine", return_value=None)
+    @patch.object(SGLangRollout, "_init_sampling_params", return_value=None)
     def test_tool_call_batch_case(self, mock_sampling, mock_engine, mock_env, mock_execute, search_rollout_config, qwen_tokenizer, qwen_model_config, search_data_proto, search_data):
         _, expect_turn_array, tool_return_array = search_data
 
@@ -285,7 +285,7 @@ class TestRolloutWithSearchTools:
         ] * 100
 
         search_rollout_config.multi_turn.max_turns = 10
-        rollout = AsyncSGLangRollout(
+        rollout = SGLangRollout(
             actor_module="",
             config=search_rollout_config,
             tokenizer=qwen_tokenizer,
@@ -327,7 +327,7 @@ class TestRolloutWithSearchTools:
             req_turns_counter[_req.batch_data_id] += 1
             return await fut
 
-        with patch.object(AsyncSGLangRollout, "_handle_engine_call", new=hacked_handle_engine_call):
+        with patch.object(SGLangRollout, "_handle_engine_call", new=hacked_handle_engine_call):
             rollout._tp_rank = 0
             loop = asyncio.get_event_loop()
             output_req_list = loop.run_until_complete(asyncio.gather(*[rollout._async_rollout_a_request(r, True, False) for r in req_list]))
