+++ verl/trainer/config/ppo_trainer.yaml
@@ -140,7 +140,7 @@ actor_rollout_ref:
       n: 1
       do_sample: False # default eager for validation
     multi_turn:
-      enable: False  # should set rollout.name to sglang_async if True
+      enable: False  # set to True for multi-turn tool interaction tasks; should set rollout.name to sglang as well
       max_turns: null  # null for no limit (default max_length // 3)
       tool_config_path: null  # null for no tool
       format: chatml  # chatml, more formats will be supported in the future
